+++ common-test/src/clojure_lsp/test_helper.clj
@@ -68,8 +68,7 @@
    (clean-db! :unit-test))
   ([env]
    (reset! db/db* (assoc db/initial-db
-                         :env env
-                         #_#_:settings {:experimental {:dep-graph-queries true}}))
+                         :env env))
    (alter-var-root #'db/current-changes-chan (constantly (async/chan 1)))
    (alter-var-root #'db/diagnostics-chan (constantly (async/chan 1)))
    (alter-var-root #'db/created-watched-files-chan (constantly (async/chan 1)))