+++ CHANGELOG.md
@@ -7,11 +7,14 @@
   - Fix extra space on :import when sorting classes. #1152
   - Bump Graalvm from 22.0.0.2 to 22.2.0 improving binary performance/size. #1154
   - Bump clj-kondo to `2022.08.03`.
+  - General release of dep-graph-queries, improving query performance. See prior release for feature details.
   
 - Editor
   - Improve completion sorting, showing locals before functions and other completion items. #1158
   - Fix hover to show current var definition docs instead of `def`/`defn`. #1157
   - Improve hover for requires to show ns docstring. #1171
+  - Fix destructuring of things that have been destructured before. #1156
+  - Some completions require that a new alias be added to the ns form. Fix this feature for Calva users, and improve performance for all users. #1068
 
 ## 2022.07.24-18.25.43